Changed implementation type from float to double, fixed matrix multiplication bug in matrix creation and general clean-up

+++ dwarfs/dense-linear-algebra/lud/c/lud_base.c
@@ -1,7 +1,7 @@
-void lud_base(float *a, int size)
+void lud_base(double *a, int size)
 {
      int i,j,k;
-     float sum;
+     double sum;
 
      for (i=0; i<size; i++){
          for (j=i; j<size; j++){
